backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1

# Scraping
scrapy==2.11.0
playwright==1.40.0
scrapy-playwright==0.0.24
beautifulsoup4==4.12.2
requests==2.31.0
fake-useragent==1.4.0
feedparser==6.0.10
yfinance==0.2.18
w3lib==2.1.2

# Scheduling and background tasks
apscheduler==3.10.4
celery==5.3.4

# AI/ML
openai==1.3.8
anthropic==0.7.7
spacy==3.7.2
transformers==4.36.0

# Data processing
pandas==2.1.4
numpy==1.25.2
python-dateutil==2.8.2

# Monitoring and logging
structlog==23.2.0
sentry-sdk==1.38.0

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Environment and configuration
python-dotenv==1.0.0